// src/app/emuns/roles.enum.ts
export enum RolesEnum {
    BcaAdmin = "bca_admin",
    OrganizationBranchManager = 'organization_branch_manager',
    OrganizationClient = 'organization_client',
    OrganizationAdmin = 'organization_admin',
    OrganizationOperatorManager = 'organization_operator_manager',
    OrganizationOperator = 'organization_operator',
    OrganizationAccountant = 'organization_accountant',           // Nouveau
    OrganizationManager = 'organization_manager',                 // Nouveau
    OrganizationHrManager = 'organization_hr_manager',            // Nouveau
    OrganizationExecutiveManager = 'organization_executive_manager', // Nouveau
    OrganizationBgo = 'organization_bgo',                         // Nouveau
    OrganizationOfficeAdmin = 'organization_office_admin',         // Nouveau
    NotFound = 'not_found'
}